Added ClassD_Full_Bridge_discrete_inverter.asc
Added a solution based on a discrete 2n2222 type inverter for driving bottom pwm
and resistor for driving top pwm input. This is simple and the signals are quite
Nice enough to get something that looks like sine wave at the output resistor be
both totempole drivers. So a speaker would produce a modulated sine wave as intended.